"install cache" setting to avoid frequent runs of pip installs...
this creates a feature to avoid running pip all the time. This does it by creating a state file inside the root of the venv to cache when the last good run of the pip command (with those arguments) was run.

For example a cache file will look like this:

$ more .nox/test-3-6/.nox_install_cache
{"[\"--upgrade\",\"-r\",\"test-requirements.txt\"]":1574446974}

There are three ways to set this

    globally on the command line: nox -install-cache 10
    globally inside the noxfile.py: nox.options.install_cache="10"
    per individual session invocations: session.install("--upgrade", "-r", "test-requirements.txt", nox_install_cache="10")