Do not merge command-line options in place
When merging options specified in the noxfile and on the command-line option, do
not use the output parameter `command_args` as the input for the merge; instead,
copy `command_args` initially and pass the copy to the merge functions.

Merge functions such as `_session_filters_merge_func` inspect `command_args` to
see if other options have been specified on the command-line. When the options
are merged in place, this check produces false positives.

For example, `nox.options.sessions` is copied into `command_args` as a part of
the merge; so it will appear to have been specified on the command-line when
merging `nox.options.pythons`, causing the latter to be ignored.