Add test fixture to generate noxfile.py
The noxfile.py is templated with the default session (`nox.options.sessions`),
and the default Python version (`nox.options.pythons`), as well as an alternate
Python version.

This allows us to avoid the situation where a test case running on one Python
version needs to launch Nox using another Python version. As a side-effect, it
makes the test cases a bit more explicit.