Decouple merging of --python with nox.options from --sessions and --keywords (#359)
* Add test fixture to generate noxfile.py

The noxfile.py is templated with the default session (`nox.options.sessions`),
and the default Python version (`nox.options.pythons`), as well as an alternate
Python version.

This allows us to avoid the situation where a test case running on one Python
version needs to launch Nox using another Python version. As a side-effect, it
makes the test cases a bit more explicit.

* Add test case for using --pythons with nox.options.sessions

* Add test case for using --sessions with nox.options.pythons

* Do not ignore nox.options.pythons when --{sessions,keywords} passed

* Do not ignore nox.options.{sessions,keywords} when --pythons passed

* Rename _{session_filters => sessions_and_keywords}_merge_func

Revert function name to the one used before --pythons was introduced.