chore(coverage): Exclude _parse_string_constant from coverage
Excluding only the Python < 3.8 branch from coverage does not work, as the CI
jobs for Python 3.6 and 3.7 will still fail. Take a pragmatic approach for now.

We do have test coverage for the main two branches of this function (not the
final `return None`), but fixing this would require combining coverage data in
CI, or using coverage-conditional-plugin.