Fix disabled reuse due to comparing sys.prefix with sys.base_prefix
When comparing the resolved interpreter to the interpreter in an existing
virtualenv, we need to use sys.base_prefix (or sys.real_prefix) for both.
The resolved interpreter is located in the virtualenv, so its sys.prefix would
point into the virtualenv instead of the base installation.

This bug results in virtualenvs never being reused.