Fix test for reusing venv environments on Windows
The test places a spurious occurrence of "virtualenv" in the pyvenv.cfg file, to
check for robust parsing. This was done by modifying the `home` key to point to
a fictitious installation at /opt/virtualenv/bin.

Unfortunately, this leads to the following error on Windows when determining the
Python base installation:

    No Python at '/opt/virtualenv/bin\python.exe'

Avoid this error by using an entirely bogus key, which should be ignored by
Python and other tools:

    bogus = virtualenv