now checkign that the cwd is unchanged
after command.run both in case of a success
or a failure (exit code being 0 or else)

We do not check that the cwd stays consistent for
the whole time the subprocess is running, hence
the behavior of the new cwd keyword argument
is undefined *during* the execution of the process.

The command.run API feels like it is atomic and not intended to be used
in exotic concurent use-cases.  This explains why no concurent tests
are added at this moment.