fix: nox.session.run-ing commands with pathlib.Path arguments
Previously this should have worked in theory, but the string
which is logged for debugging calls shlex.quote, which was
blowing up trying to quote the pathlib.Path.

I'm slightly surprised shlex.quote doesn't support pathlib.Path
objects (yet?) but for now here this just falls back to a cruder
representation when quoting fails.