feat: Added callable venv_backend
This branch allows passing a callable for `venv_backend` in `nox.session`.
This allows the user to apply any special commands during the virtualenv
creation.
As an example, `noxfile.py` includes some demo session using this functionality.

* conda-env-backend: creates an environment using `conda env create`
* conda-lock-backend: creates an environment using `conda-lock install`
* dev-example: Creates a development.  This creates an environment in
`.nox/.venv` for demo purposes

Other potential uses:

* Alter the python interpreter search path
* Add smart caching of environment.  i.e., add check in callback to
  reinstall/recreate only if dependencies have changed

The whole point is that the approach is general.

Theres a degree of buyer beware with this functionality.  In the examples
in `noxfile.py`, the correct python version must be specified in the
`environment.yaml`
files.  To me, the added functionality is worth it.